+++ Data/Conduit/Process/Unix.hs
@@ -33,7 +33,7 @@ import           Control.Exception               (Exception, SomeException,
 import           Control.Monad                   (void)
 import           Data.ByteString                 (ByteString)
 import qualified Data.ByteString.Char8           as S8
-import           Data.Conduit                    (Source, ($$))
+import           Data.Conduit                    (ConduitT, (.|), runConduit)
 import           Data.Conduit.Binary             (sinkHandle, sourceHandle)
 import qualified Data.Conduit.List               as CL
 import           Data.IORef                      (IORef, newIORef, readIORef,
@@ -61,7 +61,9 @@ import           System.Process.Internals        (ProcessHandle (..),
                                                   ProcessHandle__ (..))
 
 processHandleMVar :: ProcessHandle -> MVar ProcessHandle__
-#if MIN_VERSION_process(1, 2, 0)
+#if MIN_VERSION_process(1, 6, 0)
+processHandleMVar (ProcessHandle m _ _) = m
+#elif MIN_VERSION_process(1, 2, 0)
 processHandleMVar (ProcessHandle m _) = m
 #else
 processHandleMVar (ProcessHandle m) = m
@@ -189,7 +191,7 @@ forkExecuteLog :: ByteString -- ^ command
                -> [ByteString] -- ^ args
                -> Maybe [(ByteString, ByteString)] -- ^ environment
                -> Maybe ByteString -- ^ working directory
-               -> Maybe (Source IO ByteString) -- ^ stdin
+               -> Maybe (ConduitT () ByteString IO ()) -- ^ stdin
                -> (ByteString -> IO ()) -- ^ both stdout and stderr will be sent to this location
                -> IO ProcessHandle
 forkExecuteLog cmd args menv mwdir mstdin rlog = bracketOnError
@@ -213,6 +215,7 @@ forkExecuteLog cmd args menv mwdir mstdin rlog = bracketOnError
             , std_err = UseHandle writerH
             , close_fds = True
             , create_group = True
+            , use_process_jobs = False
 #if MIN_VERSION_process(1, 2, 0)
             , delegate_ctlc = False
 #endif
@@ -228,10 +231,10 @@ forkExecuteLog cmd args menv mwdir mstdin rlog = bracketOnError
             }
         ignoreExceptions $ addAttachMessage pipes ph
         void $ forkIO $ ignoreExceptions $
-            (sourceHandle readerH $$ CL.mapM_ rlog) `finally` hClose readerH
+            (runConduit $ sourceHandle readerH .| CL.mapM_ rlog) `finally` hClose readerH
         case (min, mstdin) of
             (Just h, Just source) -> void $ forkIO $ ignoreExceptions $
-                (source $$ sinkHandle h) `finally` hClose h
+                (runConduit $ source .| sinkHandle h) `finally` hClose h
             (Nothing, Nothing) -> return ()
             _ -> error $ "Invariant violated: Data.Conduit.Process.Unix.forkExecuteLog"
         return ph